move LazyMutable to its own file
