+++ alchemist/alchemist-cognitive-agents/src/main/kotlin/it/unibo/alchemist/model/implementations/actions/ReachKnownDestination.kt
@@ -12,6 +12,7 @@ package it.unibo.alchemist.model.implementations.actions
 import it.unibo.alchemist.model.implementations.actions.navigationstrategies.KnownDestinationReaching
 import it.unibo.alchemist.model.interfaces.NavigationAction
 import it.unibo.alchemist.model.implementations.positions.Euclidean2DPosition
+import it.unibo.alchemist.model.implementations.utils.lazyMutable
 import it.unibo.alchemist.model.interfaces.EuclideanNavigationStrategy
 import it.unibo.alchemist.model.interfaces.OrientingPedestrian
 import it.unibo.alchemist.model.interfaces.Reaction
