+++ alchemist/alchemist-cognitive-agents/src/main/kotlin/it/unibo/alchemist/model/implementations/actions/Utils.kt
@@ -40,27 +40,3 @@ fun <T, P> Sequence<P>.discardUnsuitablePositions(
          */
         environment !is PhysicsEnvironment<T, P, *, *> || environment.canNodeFitPosition(pedestrian, it)
     }
-
-/**
- * A delegate allowing to lazily initialise a mutable variable (= var).
- */
-class LazyMutable<T>(private val initializer: () -> T) : ReadWriteProperty<Any?, T> {
-
-    private var value: Optional<T> = Optional.empty()
-
-    override fun getValue(thisRef: Any?, property: KProperty<*>): T {
-        if (value.isEmpty) {
-            value = Optional.of(initializer())
-        }
-        return value.get()
-    }
-
-    override fun setValue(thisRef: Any?, property: KProperty<*>, value: T) {
-        this.value = Optional.of(value)
-    }
-}
-
-/**
- * Creates an instance of [LazyMutable] with the given [initializer].
- */
-fun <T> lazyMutable(initializer: () -> T): LazyMutable<T> = LazyMutable(initializer)
